+++ src/main/kotlin/com/codeland/uhc/phase/phases/postgame/PostgameDefault.kt
@@ -11,8 +11,8 @@ class PostgameDefault : Phase() {
 
     override fun customStart() {}
 
-    override fun updateBarLength(remainingSeconds: Int, currentTick: Int): Double {
-        return 1.0
+    override fun updateBarLength(remainingSeconds: Int, currentTick: Int): Float {
+        return 1.0f
     }
 
     override fun updateBarTitle(world: World, remainingSeconds: Int, currentTick: Int): String {
